+++ src/main/java/cam72cam/immersiverailroading/registry/EntityRollingStockDefinition.java
@@ -905,7 +905,7 @@ public abstract class EntityRollingStockDefinition {
         return tiltMultiplier;
     }
 
-    public double getBrakeShoeFriction() {
+    public float getBrakeShoeFriction() {
         return brakeCoefficient;
     }
